Rename unlock to ubuntu-core-cryptsetup and make the interface more like systemd-cryptsetup
