+++ .gitignore
@@ -5,5 +5,5 @@ examples/provision-tpm/provision-tpm
 examples/seal-key/seal-key
 examples/unseal-key/unseal-key
 examples/verify-tpm/verify-tpm
-unlock/unlock
+ubuntu-core-cryptsetup/ubuntu-core-cryptsetup
 vendor/*/
